requirements.txt
jupyterlab
cython
numpy
matplotlib
tqdm
h5py
python-dotenv
torch
poutyne
scikit-learn
tensorboard
torchsummary
